+++ src/Money.php
@@ -6,6 +6,7 @@ use Illuminate\Support\Carbon;
 use PostScripton\Money\Exceptions\MoneyHasDifferentCurrenciesException;
 use PostScripton\Money\Exceptions\NotNumericOrMoneyException;
 use PostScripton\Money\Exceptions\ServiceDoesNotSupportCurrencyException;
+use PostScripton\Money\PHPDocs\MoneyInterface;
 use PostScripton\Money\Services\ServiceInterface;
 use PostScripton\Money\Traits\MoneyHelpers;
 use PostScripton\Money\Traits\MoneyStatic;
