+++ src/MoneySettings.php
@@ -5,6 +5,7 @@ namespace PostScripton\Money;
 use PostScripton\Money\Exceptions\CurrencyDoesNotExistException;
 use PostScripton\Money\Exceptions\ShouldPublishConfigFileException;
 use PostScripton\Money\Exceptions\UndefinedOriginException;
+use PostScripton\Money\PHPDocs\MoneySettingsInterface;
 
 class MoneySettings implements MoneySettingsInterface
 {
